+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Fixed: `block-closing-brace-newline-before` no longer delivers false positives for extra semicolon.
 - Fixed: `font-family-name-quotes` is now case insensitive when hunting for font-family properties.
 - Fixed: `number-zero-length-no-unit` now ignores `deg`, `grad`, `turn` and `rad` units.
+- Fixed: improved detection of use `vertical-align` with block-level elements in `declaration-block-no-ignored-properties`.
 
 # 6.3.2